refactor(DataType): 重构所有 id 数据类型
将所有 Long 类型 id，修改为 Integer 类型

+++ src/main/java/io/github/talelin/latticy/bo/FileBO.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 /**
  * @author pedro@TaleLin
+ * @author Juzi@TaleLin
  */
 @Data
 public class FileBO {
@@ -11,7 +12,7 @@ public class FileBO {
     /**
      * 文件 id
      */
-    private Long id;
+    private Integer id;
 
     /**
      * 文件 key，上传时指定的
