+++ src/main/java/io/github/talelin/latticy/model/GroupPermissionDO.java
@@ -4,7 +4,6 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
 
@@ -19,19 +18,19 @@ public class GroupPermissionDO implements Serializable {
     private static final long serialVersionUID = -358487811336536495L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     /**
      * 分组id
      */
-    private Long groupId;
+    private Integer groupId;
 
     /**
      * 权限id
      */
-    private Long permissionId;
+    private Integer permissionId;
 
-    public GroupPermissionDO(Long groupId, Long permissionId) {
+    public GroupPermissionDO(Integer groupId, Integer permissionId) {
         this.groupId = groupId;
         this.permissionId = permissionId;
     }
