+++ src/main/java/io/github/talelin/latticy/model/UserGroupDO.java
@@ -4,7 +4,6 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
 
@@ -19,19 +18,19 @@ public class UserGroupDO implements Serializable {
     private static final long serialVersionUID = -7219009955825484511L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     /**
      * 用户id
      */
-    private Long userId;
+    private Integer userId;
 
     /**
      * 分组id
      */
-    private Long groupId;
+    private Integer groupId;
 
-    public UserGroupDO(Long userId, Long groupId) {
+    public UserGroupDO(Integer userId, Integer groupId) {
         this.userId = userId;
         this.groupId = groupId;
     }
